+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed 
+
+- Deprecate `PhoneNumber::$errorNames` in favour of `PhoneNumber::ERROR_NAMES`.
+
 ## [3.6.2] - 2022-02-24
 
 ### Fixed